Move classes to support.design.widget package..
By using the same package name as the original CollapsingToolbarLayout class, we can also access protected classes from that package, which means we don’t need to duplicaate all the helper classes.

This change moves SubtitleCollapsingTextHelper and SubtitleCollapsingToolbarLayout to android.support.design.widget, and removes all of the now-redundant helper classes.

Also removed the now unused srcDirs from the build.gradle file.
