+++ .idea/modules.xml
@@ -5,9 +5,12 @@
       <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" />
       <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" />
       <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" />
+      <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle.iml" />
       <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" />
       <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" />
       <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" />
+      <module fileurl="file://$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" filepath="$PROJECT_DIR$/collapsingtoolbarlayout-subtitle/collapsingtoolbarlayout-subtitle-collapsingtoolbarlayout-subtitle.iml" />
+      <module fileurl="file://$PROJECT_DIR$/example/example.iml" filepath="$PROJECT_DIR$/example/example.iml" />
       <module fileurl="file://$PROJECT_DIR$/example/example.iml" filepath="$PROJECT_DIR$/example/example.iml" />
       <module fileurl="file://$PROJECT_DIR$/example/example.iml" filepath="$PROJECT_DIR$/example/example.iml" />
       <module fileurl="file://$PROJECT_DIR$/example/example.iml" filepath="$PROJECT_DIR$/example/example.iml" />
