+++ example/src/main/java/com/example/collapsingtoolbarlayoutsubtitle/SubtitleActivity.java
@@ -4,7 +4,7 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
-import com.hendraanggrian.widget.SubtitleCollapsingToolbarLayout;
+import android.support.design.widget.SubtitleCollapsingToolbarLayout;
 
 /**
  * @author Hendra Anggrian (hendraanggrian@gmail.com)
